fix!: remove client base module and client macro (#110)
check #97 and #28
