+++ lib/mix/interactive/supervised_shell.ex
@@ -207,7 +207,7 @@ defmodule Mix.Interactive.SupervisedShell do
   defp start_client(%{client_opts: opts}) do
     IO.puts("#{UI.colors().info}• Starting client...#{UI.colors().reset}")
 
-    case Anubis.Client.Base.start_link(opts) do
+    case Anubis.Client.start_link_server(opts) do
       {:ok, pid} ->
         IO.puts("#{UI.colors().success}✓ Client started#{UI.colors().reset}")
         {:ok, pid}
