+++ priv/dev/client/lib/incomer/application.ex
@@ -6,7 +6,12 @@ defmodule Incomer.Application do
   @impl true
   def start(_, _) do
     children = [
-      {Incomer.Client, transport: {:streamable_http, base_url: "http://localhost:4000"}}
+      {Anubis.Client,
+       name: Incomer.Client,
+       transport: {:streamable_http, base_url: "http://localhost:4000"},
+       client_info: %{"name" => "incomer", "version" => "0.1.0"},
+       capabilities: %{},
+       protocol_version: "2025-03-26"}
     ]
 
     Supervisor.start_link(children, strategy: :one_for_one, name: Incomer.Supervisor)
