+++ priv/dev/client/lib/incomer/client.ex
@@ -1,5 +1,3 @@
 defmodule Incomer.Client do
   @moduledoc false
-
-  use Anubis.Client, name: "incomer", version: "0.1.0", protocol_version: "2025-03-26"
 end
